+++ standalone/include/psen_scan_v2_standalone/protocol_layer/scanner_state_machine_def.h
@@ -23,6 +23,7 @@ namespace protocol_layer
 inline ScannerProtocolDef::ScannerProtocolDef(const ScannerConfiguration config,
                                               const communication_layer::NewMessageCallback& control_msg_callback,
                                               const communication_layer::ErrorCallback& control_error_callback,
+                                              const communication_layer::ErrorCallback& start_error_callback,
                                               const communication_layer::NewMessageCallback& data_msg_callback,
                                               const communication_layer::ErrorCallback& data_error_callback,
                                               const ScannerStartedCallback& scanner_started_callback,
@@ -43,6 +44,7 @@ inline ScannerProtocolDef::ScannerProtocolDef(const ScannerConfiguration config,
                  config_.scannerDataPort())
   , scanner_started_callback_(scanner_started_callback)
   , scanner_stopped_callback_(scanner_stopped_callback)
+  , start_error_callback_(start_error_callback)
   , inform_user_about_laser_scan_callback_(laser_scan_callback)
   , start_timeout_callback_(start_timeout_callback)
   , monitoring_frame_timeout_callback_(monitoring_frame_timeout_callback)
@@ -108,7 +110,6 @@ void ScannerProtocolDef::WaitForMonitoringFrame::on_entry(Event const&, FSM& fsm
   // Start watchdog...
   fsm.monitoring_frame_watchdog_ =
       fsm.watchdog_factory_.create(WATCHDOG_TIMEOUT, fsm.monitoring_frame_timeout_callback_);
-  fsm.scanner_started_callback_();
 }
 
 template <class Event, class FSM>
@@ -123,10 +124,12 @@ template <class Event, class FSM>
 void ScannerProtocolDef::Stopped::on_entry(Event const&, FSM& fsm)
 {
   PSENSCAN_DEBUG("StateMachine", "Entering state: Stopped");
-  fsm.scanner_stopped_callback_();
 }
 
 DEFAULT_ON_EXIT_IMPL(Stopped)
+
+DEFAULT_ON_ENTRY_IMPL(Error)
+DEFAULT_ON_EXIT_IMPL(Error)
 
 // \endcond
 //+++++++++++++++++++++++++++++++++ Actions +++++++++++++++++++++++++++++++++++
@@ -182,6 +185,29 @@ inline void ScannerProtocolDef::handleMonitoringFrame(const scanner_events::RawM
     PSENSCAN_ERROR("StateMachine", e.what());
   }
   // LCOV_EXCL_STOP
+}
+
+inline void ScannerProtocolDef::notifyUserAboutStart(scanner_events::RawReplyReceived const& reply_event)
+{
+  scanner_started_callback_();
+}
+
+inline void ScannerProtocolDef::notifyUserAboutStop(scanner_events::RawReplyReceived const& reply_event)
+{
+  scanner_stopped_callback_();
+}
+
+inline void ScannerProtocolDef::notifyUserAboutUnknownStartReply(scanner_events::RawReplyReceived const& reply_event)
+{
+  const data_conversion_layer::scanner_reply::Message msg{ data_conversion_layer::scanner_reply::deserialize(
+      *(reply_event.data_)) };
+  start_error_callback_(
+      fmt::format("Unknown result code {:#04x} in start reply.", static_cast<uint32_t>(msg.result())));
+}
+
+inline void ScannerProtocolDef::notifyUserAboutRefusedStartReply(scanner_events::RawReplyReceived const& reply_event)
+{
+  start_error_callback_("Request refused by device.");
 }
 
 inline void ScannerProtocolDef::checkForDiagnosticErrors(const data_conversion_layer::monitoring_frame::Message& msg)
@@ -284,11 +310,44 @@ inline void ScannerProtocolDef::checkForInternalErrors(const data_conversion_lay
   // LCOV_EXCL_STOP
 }
 
-inline bool ScannerProtocolDef::isStartReply(scanner_events::RawReplyReceived const& reply_event)
+inline bool ScannerProtocolDef::isAcceptedStartReply(scanner_events::RawReplyReceived const& reply_event)
 {
   const data_conversion_layer::scanner_reply::Message msg{ data_conversion_layer::scanner_reply::deserialize(
       *(reply_event.data_)) };
-  checkForInternalErrors(msg);
+  return isStartReply(msg) && isAcceptedReply(msg);
+}
+
+inline bool ScannerProtocolDef::isUnknownStartReply(scanner_events::RawReplyReceived const& reply_event)
+{
+  const data_conversion_layer::scanner_reply::Message msg{ data_conversion_layer::scanner_reply::deserialize(
+      *(reply_event.data_)) };
+  return isStartReply(msg) && isUnknownReply(msg);
+}
+
+inline bool ScannerProtocolDef::isRefusedStartReply(scanner_events::RawReplyReceived const& reply_event)
+{
+  const data_conversion_layer::scanner_reply::Message msg{ data_conversion_layer::scanner_reply::deserialize(
+      *(reply_event.data_)) };
+  return isStartReply(msg) && isRefusedReply(msg);
+}
+
+inline bool ScannerProtocolDef::isAcceptedReply(data_conversion_layer::scanner_reply::Message const& msg)
+{
+  return msg.result() == data_conversion_layer::scanner_reply::Message::OperationResult::accepted;
+}
+
+inline bool ScannerProtocolDef::isUnknownReply(data_conversion_layer::scanner_reply::Message const& msg)
+{
+  return msg.result() == data_conversion_layer::scanner_reply::Message::OperationResult::unknown;
+}
+
+inline bool ScannerProtocolDef::isRefusedReply(data_conversion_layer::scanner_reply::Message const& msg)
+{
+  return msg.result() == data_conversion_layer::scanner_reply::Message::OperationResult::refused;
+}
+
+inline bool ScannerProtocolDef::isStartReply(data_conversion_layer::scanner_reply::Message const& msg)
+{
   return msg.type() == data_conversion_layer::scanner_reply::Message::Type::start;
 }