+++ standalone/test/integration_tests/api/integrationtest_scanner_api.cpp
@@ -227,6 +227,40 @@ TEST_F(ScannerAPITests, shouldReturnInvalidFutureWhenStartIsCalledSecondTime)
   EXPECT_SCANNER_TO_STOP_SUCCESSFULLY(hw_mock_, driver_);
 }
 
+TEST_F(ScannerAPITests, startShouldReturnFutureWithExceptionIfStartRequestRefused)
+{
+  setUpScannerConfig();
+  setUpScannerV2Driver();
+  setUpScannerHwMock();
+
+  util::Barrier start_req_received_barrier;
+  EXPECT_START_REQUEST_CALL(*hw_mock_, *config_).WillOnce(OpenBarrier(&start_req_received_barrier));
+
+  std::future<void> start_future = driver_->start();
+  start_req_received_barrier.waitTillRelease(DEFAULT_TIMEOUT);
+
+  hw_mock_->sendStartReply(data_conversion_layer::scanner_reply::Message::OperationResult::refused);
+  EXPECT_FUTURE_IS_READY(start_future, DEFAULT_TIMEOUT);
+  EXPECT_THROW(start_future.get(), std::runtime_error);
+}
+
+TEST_F(ScannerAPITests, startShouldReturnFutureWithExceptionIfUnknownResultSent)
+{
+  setUpScannerConfig();
+  setUpScannerV2Driver();
+  setUpScannerHwMock();
+
+  util::Barrier start_req_received_barrier;
+  EXPECT_START_REQUEST_CALL(*hw_mock_, *config_).WillOnce(OpenBarrier(&start_req_received_barrier));
+
+  std::future<void> start_future = driver_->start();
+  start_req_received_barrier.waitTillRelease(DEFAULT_TIMEOUT);
+
+  hw_mock_->sendStartReply(data_conversion_layer::scanner_reply::Message::OperationResult::unknown);
+  EXPECT_FUTURE_IS_READY(start_future, DEFAULT_TIMEOUT);
+  EXPECT_THROW(start_future.get(), std::runtime_error);
+}
+
 TEST_F(ScannerAPITests, startShouldSucceedDespiteUnexpectedMonitoringFrame)
 {
   setUpScannerConfig();