+++ test/integration_tests/integrationtest_ros_scanner_node.cpp
@@ -15,7 +15,6 @@
 
 #include <memory>
 #include <chrono>
-#include <thread>
 #include <future>
 
 #include <gtest/gtest.h>